Implement abs_compare() and add equal_with_nan_inf()
Tests now pass
